backend/requirements.txt
# Web Framework
fastapi==0.111.0
uvicorn[standard]==0.30.1

# Database
sqlalchemy[asyncio]==2.0.41
aiosqlite==0.20.0
asyncpg==0.30.0
alembic==1.13.1

# HTTP Client & Scraping
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.2
Pillow>=10.4.0

# Scheduler
apscheduler==3.10.4

# Auth / Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Config
pydantic-settings==2.2.1
python-dotenv==1.0.1
# orjson removed — incompatible with Python 3.13; FastAPI uses stdlib json by default
